refactor: `UpdateNote`
